# docs/specs/vehicle-listings/openapi.yaml
openapi: 3.0.3
info:
  title: ZoomLoop Vehicle API
  description: API for searching and managing vehicle listings on ZoomLoop
  version: 1.0.0
  contact:
    name: ZoomLoop Support
  license:
    name: MIT

servers:
  - url: https://localhost:7217
    description: Development server
  - url: https://api.zoomloop.com
    description: Production server

tags:
  - name: Vehicles
    description: Vehicle search and management operations

paths:
  /api/vehicle/search:
    post:
      operationId: searchVehicles
      summary: Search vehicles with filters and pagination
      description: |
        Search for vehicles using various filters including make, model, price range,
        year, transmission, and more. Supports server-side pagination and sorting.
        This endpoint is publicly accessible without authentication.
      tags:
        - Vehicles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchVehiclesRequest'
            examples:
              basicSearch:
                summary: Basic search
                value:
                  page: 1
                  pageSize: 12
              filterByMake:
                summary: Filter by make
                value:
                  filters:
                    makes: ["Honda", "Toyota"]
                  page: 1
                  pageSize: 12
              priceRange:
                summary: Filter by price range
                value:
                  filters:
                    price:
                      min: 20000
                      max: 40000
                  page: 1
                  pageSize: 12
              sortByPrice:
                summary: Sort by price ascending
                value:
                  sort:
                    - field: price
                      direction: asc
                  page: 1
                  pageSize: 12
      responses:
        '200':
          description: Successful search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchVehiclesResponse'
        '400':
          description: Bad request - invalid filter parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /api/vehicle/{vehicleId}:
    get:
      operationId: getVehicleById
      summary: Get vehicle by ID
      description: Retrieve detailed information about a specific vehicle
      tags:
        - Vehicles
      security:
        - bearerAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique vehicle identifier
      responses:
        '200':
          description: Vehicle details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetVehicleByIdResponse'
        '404':
          description: Vehicle not found
          content:
            application/json:
              schema:
                type: string
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /api/vehicle:
    get:
      operationId: getVehicles
      summary: Get all vehicles
      description: Retrieve all vehicles (paginated internally)
      tags:
        - Vehicles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetVehiclesResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

    post:
      operationId: createVehicle
      summary: Create a new vehicle listing
      description: Create a new vehicle in the system
      tags:
        - Vehicles
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleRequest'
      responses:
        '200':
          description: Vehicle created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateVehicleResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /api/vehicle/page/{pageSize}/{index}:
    get:
      operationId: getVehiclesPage
      summary: Get paginated vehicles
      description: Retrieve vehicles with pagination
      tags:
        - Vehicles
      security:
        - bearerAuth: []
      parameters:
        - name: pageSize
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Number of items per page
        - name: index
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: Page number (1-based)
      responses:
        '200':
          description: Paginated vehicle list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetVehiclesPageResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /api/vehicle/suggestions:
    get:
      operationId: getVehicleSuggestions
      summary: Get vehicle suggestions
      description: Get autocomplete suggestions for vehicle search
      tags:
        - Vehicles
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Search query for suggestions
      responses:
        '200':
          description: Vehicle suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetVehicleSuggestionsResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SearchVehiclesRequest:
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/SearchFilters'
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortCriteria'
        page:
          type: integer
          minimum: 1
          default: 1
          description: Page number (1-based)
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
          description: Number of items per page

    SearchFilters:
      type: object
      properties:
        isNewest:
          type: boolean
          description: Filter for newest/featured vehicles
        colors:
          type: array
          items:
            type: string
          description: Filter by exterior colors
        interiorCondition:
          type: string
          description: Filter by interior condition
        exteriorCondition:
          type: string
          description: Filter by exterior condition
        doors:
          type: integer
          description: Filter by number of doors
        transmissions:
          type: array
          items:
            type: string
            enum: [automatic, manual]
          description: Filter by transmission types
        price:
          $ref: '#/components/schemas/PriceRange'
        age:
          $ref: '#/components/schemas/IntRange'
        year:
          $ref: '#/components/schemas/IntRange'
        makes:
          type: array
          items:
            type: string
          description: Filter by vehicle makes
        models:
          type: array
          items:
            type: string
          description: Filter by vehicle models
        accidentFree:
          type: boolean
          description: Filter for accident-free vehicles only
        kilometers:
          $ref: '#/components/schemas/IntRange'

    PriceRange:
      type: object
      properties:
        min:
          type: number
          description: Minimum price
        max:
          type: number
          description: Maximum price

    IntRange:
      type: object
      properties:
        min:
          type: integer
          description: Minimum value
        max:
          type: integer
          description: Maximum value

    SortCriteria:
      type: object
      required:
        - field
      properties:
        field:
          type: string
          enum: [year, mileage, price, make, model, color, doors, transmission, age, listeddate]
          description: Field to sort by
        direction:
          type: string
          enum: [asc, desc]
          default: asc
          description: Sort direction

    SearchVehiclesResponse:
      type: object
      required:
        - items
        - total
        - page
        - pageSize
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VehicleSearchResultDto'
        total:
          type: integer
          description: Total number of matching vehicles
        page:
          type: integer
          description: Current page number
        pageSize:
          type: integer
          description: Number of items per page
        appliedFilters:
          $ref: '#/components/schemas/SearchFilters'

    VehicleSearchResultDto:
      type: object
      required:
        - vehicleId
        - make
        - model
        - year
      properties:
        vehicleId:
          type: string
          format: uuid
          description: Unique vehicle identifier
        vin:
          type: string
          description: Vehicle Identification Number
        stockNumber:
          type: string
          description: Dealer stock number
        make:
          type: string
          description: Vehicle manufacturer
        model:
          type: string
          description: Vehicle model name
        year:
          type: integer
          description: Model year
        trim:
          type: string
          description: Trim level
        mileage:
          type: integer
          description: Odometer reading in kilometers
        exteriorColor:
          type: string
          description: Exterior color
        interiorColor:
          type: string
          description: Interior color
        transmission:
          type: string
          enum: [automatic, manual]
          description: Transmission type
        doors:
          type: integer
          description: Number of doors
        price:
          type: number
          nullable: true
          description: Listing price
        isNew:
          type: boolean
          description: Whether the vehicle is new
        isCertified:
          type: boolean
          description: Whether the vehicle is certified pre-owned
        accidentFree:
          type: boolean
          nullable: true
          description: Whether the vehicle is accident-free
        primaryImageUrl:
          type: string
          format: uri
          nullable: true
          description: URL of the primary vehicle image
        listedDate:
          type: string
          format: date-time
          nullable: true
          description: Date when the vehicle was listed

    GetVehicleByIdResponse:
      type: object
      properties:
        vehicle:
          $ref: '#/components/schemas/VehicleDto'

    VehicleDto:
      type: object
      properties:
        vehicleId:
          type: string
          format: uuid
        vin:
          type: string
        stockNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        trim:
          type: string
        mileage:
          type: integer
        exteriorColor:
          type: string
        interiorColor:
          type: string
        transmission:
          type: string
        fuelType:
          type: string
        driveType:
          type: string
        bodyType:
          type: string
        doors:
          type: integer
        seats:
          type: integer
        engineSize:
          type: string
        cylinders:
          type: integer
        horsepower:
          type: integer
        price:
          type: number
        isNew:
          type: boolean
        isCertified:
          type: boolean
        images:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string

    GetVehiclesResponse:
      type: object
      properties:
        vehicles:
          type: array
          items:
            $ref: '#/components/schemas/VehicleDto'

    GetVehiclesPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VehicleDto'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    CreateVehicleRequest:
      type: object
      required:
        - make
        - model
        - year
      properties:
        vin:
          type: string
        stockNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        trim:
          type: string
        mileage:
          type: integer
        exteriorColor:
          type: string
        interiorColor:
          type: string
        transmission:
          type: string
        fuelType:
          type: string
        driveType:
          type: string
        price:
          type: number
        isNew:
          type: boolean
        isCertified:
          type: boolean

    CreateVehicleResponse:
      type: object
      properties:
        vehicle:
          $ref: '#/components/schemas/VehicleDto'

    GetVehicleSuggestionsResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: string

    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string